move feedback classes to toolkit-lib module.
