+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib-java/build.gradle
@@ -1,3 +1,7 @@
+plugins {
+    id "org.jetbrains.kotlin.jvm"
+}
+
 intellij {
     plugins = ['java', 'maven', 'maven-model', 'gradle']
 }
