+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/feedback/NewCustomerIssueFeedbackAction.java
@@ -6,15 +6,14 @@
 package com.microsoft.azure.toolkit.intellij.feedback;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.microsoft.azuretools.telemetry.TelemetryConstants;
 import com.microsfot.azure.toolkit.intellij.feedback.GithubIssue;
 import com.microsfot.azure.toolkit.intellij.feedback.NewGithubIssueAction;
-import com.microsfot.azure.toolkit.intellij.feedback.ReportableFeatureRequest;
+import com.microsfot.azure.toolkit.intellij.feedback.ReportableIssue;
+import com.microsoft.azuretools.telemetry.TelemetryConstants;
 
-public class NewFeatureRequestFeedbackAction extends NewGithubIssueAction {
-    public NewFeatureRequestFeedbackAction() {
-        super(new GithubIssue<>(new ReportableFeatureRequest("Feature Request"))
-            .withLabel("feature-request"), "Request Features");
+public class NewCustomerIssueFeedbackAction extends NewGithubIssueAction {
+    public NewCustomerIssueFeedbackAction() {
+        super(new GithubIssue<>(new ReportableIssue("Customer Issues")), "Report Issues");
     }
 
     @Override
