+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib-java/src/main/kotlin/com/microsfot/azure/toolkit/intellij/feedback/ReportableIssue.kt
