+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib-java/src/main/resources/META-INF/azure-intellij-plugin-lib-java.xml
@@ -8,10 +8,18 @@
 
     <actions>
         <action id="Actions.WhatsNew"
-                class="com.microsoft.azure.toolkit.intellij.common.action.WhatsNewAction" text="What's New in Azure Toolkit"
+                class="com.microsoft.azure.toolkit.intellij.common.action.WhatsNewAction"
+                text="What's New in Azure Toolkit"
                 description="Show new features of Azure Toolkit for Java"/>
+        <action id="AzureToolkit.GithubIssue"
+                class="com.microsoft.azure.toolkit.intellij.feedback.NewCustomerIssueFeedbackAction"
+                text="Report an Issue"/>
+        <action id="AzureToolkit.FeatureRequest"
+                class="com.microsoft.azure.toolkit.intellij.feedback.NewFeatureRequestFeedbackAction"
+                text="Request a Feature"/>
     </actions>
     <extensions defaultExtensionNs="com.intellij">
         <postStartupActivity implementation="com.microsoft.azure.toolkit.intellij.common.action.WhatsNewStartupActivity"/>
+        <errorHandler implementation="com.microsfot.azure.toolkit.intellij.feedback.MSErrorReportHandler"/>
     </extensions>
 </idea-plugin>
